speechProcessingService/requirements.txt
fastapi
uvicorn
httpx
python-multipart
faster-whisper
pydantic
python-dotenv
torch
transformers
git+https://github.com/huggingface/parler-tts.git
indic-nlp-library
sacremoses
sentencepiece
TTS
edge-tts
